feat!: upgrade quinn to 0.8
Use the new version of the `quinn` crate, which has a different API. Constructing and configuring
endpoints was changed, with some of the functionality even moving to the `rustls` crate, which is
also upgraded here to the latest version.

BREAKING CHANGE: the API for `ServerEndpoint` and `ClientEndpoint` have both changed to construct
instances by providing paths to certificates and keys, rather than a directory containing them. To
me, for a generic library, it makes more sense for the caller to provide certs and keys from
wherever they wanted to and not constrain them to using DER certificates with a particular file
name. This change also places responsibility on the caller to generate self-signed certificates,
which I would argue should probably be done in the Safe CLI and Auth processes, not even in
`sn_api`.

Some additional functionality was added to validate the provided certificate.

Unit tests were added for both these structs where possible, though things were made difficult by
the `quinn` and `rustls` crates, since they hide information after it's been set. For more info, see
my doc comments on `ClientEndpoint`.

I added the `ping` example to CI for integration testing the code that requires a real connection.
It may have been possible to unit test this code by creating a trait to implement a fake connection,
but I decided that was excessive for a small library.